[#374] Refactor oversized modules and duplicate utilities
