--- src/utils/listAdminUsers.ts
@@ -1,29 +1,0 @@
-import { getFunctions, httpsCallable } from "firebase/functions";
-import { firebaseApp } from "../config/firebase";
-import type { AdminUser } from "../types";
-
-export type { AdminUser };
-
-interface ListAdminUsersResponse {
-  users: AdminUser[];
-}
-
-/**
- * Lists all users from Firebase Auth that have the admin claim
- * @returns Promise with the list of admin users
- */
-export async function listAdminUsers(): Promise<{ success: boolean; users?: AdminUser[]; error?: string }> {
-  try {
-    const functions = getFunctions(firebaseApp, "europe-west2");
-    const listAdminUsersCallable = httpsCallable<ListAdminUsersResponse>(functions, "listAdminUsers");
-    
-    const result = await listAdminUsersCallable();
-    return { success: true, users: (result.data as ListAdminUsersResponse).users };
-  } catch (error: any) {
-    return { 
-      success: false, 
-      error: error?.message || "Failed to list admin users" 
-    };
-  }
-}
-
